Changed some indexes and removed redundies
