+++ src/Services/Event/Event.API/Mapper/EventProfile.cs
@@ -5,33 +5,12 @@ using AutoMapper;
 using EMS.Events;
 using EMS.Event_Services.API.Context.Model;
 using EMS.Event_Services.API.Controllers.Request;
-
+using EMS.TemplateWebHost.Customization;
 
 
 namespace EMS.Event_Services.API.Mapper
 {
-    public static class MyExtensions
-    {
-        public static IMappingExpression<TSource, TDestination> Transform<TSource, TDestination, TSourceMember>(
-            this IMappingExpression<TSource, TDestination> map,
-            Expression<Func<TDestination, object>> selector,
-            Expression<Func<TSource, TSourceMember>> transform
-        )
-        {
-            map.ForMember(selector, config =>
-                config.MapFrom(transform));
-            return map;
-        }
-    }
 
-    public class RoomEventFormatter : IValueConverter<Guid, RoomEvent>
-    {
-        public RoomEvent Convert(Guid source, ResolutionContext context)
-            => new RoomEvent()
-            {
-                RoomId = source
-            };
-    }
 
     public class EventProfile : Profile
     {
